+++ tests/gepol/gepol_CH3+_Cs.cpp
@@ -65,7 +65,7 @@ struct GePolCavityCsAddTest {
 BOOST_FIXTURE_TEST_CASE(size, GePolCavityCsAddTest)
 {
     int size = 384;
-    int actualSize = cavity.size();
+    size_t actualSize = cavity.size();
     BOOST_REQUIRE_EQUAL(size, actualSize);
 }
 
@@ -98,7 +98,7 @@ BOOST_FIXTURE_TEST_CASE(volume, GePolCavityCsAddTest)
     Eigen::Matrix3Xd elementCenter = cavity.elementCenter();
     Eigen::Matrix3Xd elementNormal = cavity.elementNormal();
     double actualVolume = 0;
-    for ( int i = 0; i < cavity.size(); ++i ) {
+    for ( size_t i = 0; i < cavity.size(); ++i ) {
         actualVolume += cavity.elementArea(i) * elementCenter.col(i).dot(elementNormal.col(
                             i));
     }
